Refactor to cleaner structure
