+++ lib/isodoc-yaml/i18n-fr.yaml
