+++ lib/isodoc/function.rb
@@ -1,0 +1,4 @@
+module IsoDoc
+  module Function
+  end
+end
