+++ lib/isodoc/function/lists.rb
@@ -1,5 +1,6 @@
-module IsoDoc
-  class Common
+module IsoDoc::Function
+  module Lists
+
     def ul_parse(node, out)
       out.ul do |ul|
         node.children.each { |n| parse(n, ul) }
@@ -21,8 +22,8 @@ module IsoDoc
 
     # We don't really want users to specify type of ordered list;
     # we will use a fixed hierarchy as practiced by ISO (though not
-    # fully spelled out): a) 1) i) A) I) 
-   
+    # fully spelled out): a) 1) i) A) I)
+
     def ol_depth(node)
       depth = node.ancestors("ul, ol").size + 1
       type = :alphabet
