+++ lib/isodoc/html_convert.rb
@@ -1,0 +1,12 @@
+require_relative "html_function/comments.rb"
+require_relative "html_function/footnotes.rb"
+require_relative "html_function/html.rb"
+
+module IsoDoc
+  class HtmlConvert < ::IsoDoc::Convert
+
+    include HtmlFunction::Comments
+    include HtmlFunction::Footnotes
+    include HtmlFunction::Html
+  end
+end
