+++ lib/isodoc/html_function.rb
@@ -1,0 +1,4 @@
+module IsoDoc
+  module HtmlFunction
+  end
+end
