--- lib/isodoc/htmlconvert/convert.rb
@@ -1,13 +1,0 @@
-require "uuidtools"
-require "html2doc"
-require "liquid"
-
-require_relative "comments.rb"
-require_relative "footnotes.rb"
-require_relative "html.rb"
-
-module IsoDoc
-
-    class Convert < Common
-    end 
-end
