+++ lib/isodoc/word_function.rb
@@ -1,0 +1,4 @@
+module IsoDoc
+  module WordFunction
+  end
+end
